Added exception handling to the router service repository files for all endpoints
