+++ src/routers/note_router.py
@@ -32,8 +32,8 @@ router = APIRouter()
     status_code=status.HTTP_200_OK,
 )
 async def get_all_notes(
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method is to get all notes in the database, where deleted field is set to be 0.
@@ -61,11 +61,11 @@ async def get_all_notes(
     status_code=status.HTTP_200_OK,
 )
 async def get_note_by_id(
-    note_id: int,
-    response: Response,
-    if_none_match: str | None = Header(default=None),
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        response: Response,
+        if_none_match: str | None = Header(default=None),
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method to get a note by its id.
@@ -101,9 +101,9 @@ async def get_note_by_id(
     status_code=status.HTTP_200_OK,
 )
 async def get_users_notes(
-    user_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        user_id: int,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This endpoint get all the user's notes available in the database with deleted field set to 0.
@@ -131,9 +131,9 @@ async def get_users_notes(
     status_code=status.HTTP_200_OK,
 )
 async def get_note_history(
-    note_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This endpoint to get the history and all the previous versions of a certain note.
@@ -159,10 +159,10 @@ async def get_note_history(
     status_code=status.HTTP_200_OK,
 )
 async def get_note_old_version(
-    note_id: int,
-    history_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        history_id: int,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This endpoint returns a certain version of a note from its history.
@@ -187,9 +187,9 @@ async def get_note_old_version(
     status_code=status.HTTP_201_CREATED,
 )
 async def add_new_note(
-    note: NoteIn,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note: NoteIn,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method adds new note to the database.
@@ -199,9 +199,13 @@ async def add_new_note(
     :param session: This is the async session used to handle the database.
     :return: The note added.
     """
-    note_service = NoteService(NoteRepository(session))
-    note = await note_service.add_new_note(note)
-    return note
+    try:
+        note_service = NoteService(NoteRepository(session))
+        note = await note_service.add_new_note(note)
+        return note
+    except Exception as e:
+        await session.rollback()
+        raise e
 
 
 @router.patch(
@@ -217,10 +221,10 @@ async def add_new_note(
     status_code=status.HTTP_200_OK,
 )
 async def update_note(
-    note_id: int,
-    note: NoteUpdate,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        note: NoteUpdate,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This endpoint to update available note's data, the note shall be available if not HTTPException 404 is raised.
@@ -231,9 +235,13 @@ async def update_note(
     :param session: This is the async session used to handle the database.
     :return: The updated note.
     """
-    note_service = NoteService(NoteRepository(session))
-    note = await note_service.update_note(note_id, note)
-    return note
+    try:
+        note_service = NoteService(NoteRepository(session))
+        note = await note_service.update_note(note_id, note)
+        return note
+    except Exception as e:
+        await session.rollback()
+        raise e
 
 
 @router.delete(
@@ -248,9 +256,9 @@ async def update_note(
     status_code=status.HTTP_200_OK,
 )
 async def delete_note(
-    note_id: int,
-    user: User = Depends(check_token),
-    session: AsyncSession = Depends(Connection.get_session),
+        note_id: int,
+        user: User = Depends(check_token),
+        session: AsyncSession = Depends(Connection.get_session),
 ):
     """
     This method to delete a note from the database if available, else it raises a 404 HTTPException.
@@ -260,5 +268,9 @@ async def delete_note(
     :param session: This is the async session used to handle the database.
     :return: Successful message.
     """
-    note_service = NoteService(NoteRepository(session))
-    return await note_service.delete_note(note_id)
+    try:
+        note_service = NoteService(NoteRepository(session))
+        return await note_service.delete_note(note_id)
+    except Exception as e:
+        await session.rollback()
+        raise e
