+++ src/services/note_service.py
@@ -24,21 +24,24 @@ class NoteService:
 
         :return: The returned value is a list of notes if found.
         """
-        notes: list[Note] | None = await self.note_repository.get_all_notes()
-        if not notes:
-            raise HTTPException(status_code=404, detail="No notes are found")
+        try:
+            notes: list[Note] | None = await self.note_repository.get_all_notes()
+            if not notes:
+                raise HTTPException(status_code=404, detail="No notes are found")
 
-        return [
-            NoteOut(
-                id=note.id,
-                title=note.title,
-                content=note.content,
-                username=note.user.username,
-                parent=ParentOut(id=note.parent.id, name=note.parent.name),
-                tags=[TagOut(id=tag.id, name=tag.name) for tag in note.tags],
-            )
-            for note in notes
-        ]
+            return [
+                NoteOut(
+                    id=note.id,
+                    title=note.title,
+                    content=note.content,
+                    username=note.user.username,
+                    parent=ParentOut(id=note.parent.id, name=note.parent.name),
+                    tags=[TagOut(id=tag.id, name=tag.name) for tag in note.tags],
+                )
+                for note in notes
+            ]
+        except Exception as e:
+            raise e
 
     async def get_note_by_note_id(self, note_id: int) -> NoteOut | None:
         """
@@ -47,19 +50,22 @@ class NoteService:
         :param note_id: The id of the note to be found.
         :return: The note's data.
         """
-        note: Note | None = await self.note_repository.get_note_by_id(note_id)
-        if not note:
-            raise HTTPException(status_code=404, detail=f"Note {note_id} not found")
+        try:
+            note: Note | None = await self.note_repository.get_note_by_id(note_id)
+            if not note:
+                raise HTTPException(status_code=404, detail=f"Note {note_id} not found")
 
-        note_out = NoteOut(
-            id=note.id,
-            title=note.title,
-            content=note.content,
-            username=note.user.username,
-            parent=ParentOut(id=note.parent.id, name=note.parent.name),
-            tags=[TagOut(id=tag.id, name=tag.name) for tag in note.tags],
-        )
-        return note_out
+            note_out = NoteOut(
+                id=note.id,
+                title=note.title,
+                content=note.content,
+                username=note.user.username,
+                parent=ParentOut(id=note.parent.id, name=note.parent.name),
+                tags=[TagOut(id=tag.id, name=tag.name) for tag in note.tags],
+            )
+            return note_out
+        except Exception as e:
+            raise e
 
     async def update_note(self, note_id: int, note: NoteUpdate) -> NoteOut:
         """
@@ -70,22 +76,24 @@ class NoteService:
         :param note: New note's data to update.
         :return: The update note, if not found it raised 404 HTTPException.
         """
+        try:
+            stored_note = await self.note_repository.get_note_by_id(note_id)
 
-        stored_note = await self.note_repository.get_note_by_id(note_id)
+            if not stored_note:
+                raise HTTPException(status_code=404, detail=f"Note {note_id} not found")
 
-        if not stored_note:
-            raise HTTPException(status_code=404, detail=f"Note {note_id} not found")
+            await self.note_repository.update_note(stored_note, note)
 
-        await self.note_repository.update_note(stored_note, note)
-
-        return NoteOut(
-            id=stored_note.id,
-            title=stored_note.title,
-            content=stored_note.content,
-            username=stored_note.user.username,
-            parent=ParentOut(id=stored_note.parent.id, name=stored_note.parent.name),
-            tags=[TagOut(id=tag.id, name=tag.name) for tag in stored_note.tags],
-        )
+            return NoteOut(
+                id=stored_note.id,
+                title=stored_note.title,
+                content=stored_note.content,
+                username=stored_note.user.username,
+                parent=ParentOut(id=stored_note.parent.id, name=stored_note.parent.name),
+                tags=[TagOut(id=tag.id, name=tag.name) for tag in stored_note.tags],
+            )
+        except Exception as e:
+            raise e
 
     async def delete_note(self, note_id: int):
         """
@@ -95,8 +103,11 @@ class NoteService:
         :param note_id: The id of the note to be deleted.
         :return: True on Success, else it raised 404 HTTPException.
         """
-        await self.note_repository.delete_note(note_id)
-        return True
+        try:
+            await self.note_repository.delete_note(note_id)
+            return True
+        except Exception as e:
+            raise e
 
     async def add_new_note(self, note: NoteIn):
         """
@@ -105,25 +116,27 @@ class NoteService:
         :param note: The new note information.
         :return: The new note created.
         """
+        try:
+            new_note = Note(
+                title=note.title,
+                content=note.content,
+                user_id=note.user_id,
+                parent_id=note.parent_id,
+            )
 
-        new_note = Note(
-            title=note.title,
-            content=note.content,
-            user_id=note.user_id,
-            parent_id=note.parent_id,
-        )
+            await self.note_repository.add_new_note(new_note)
 
-        await self.note_repository.add_new_note(new_note)
+            tags = note.tag_ids
+            if tags:
+                for tag in tags:
+                    await self.note_repository.add_tag_note(new_note.id, tag)
 
-        tags = note.tag_ids
-        if tags:
-            for tag in tags:
-                await self.note_repository.add_tag_note(new_note.id, tag)
-
-        return {
-            "details": "note is added successfully",
-            "note": {"id": new_note.id, "title": new_note.title},
-        }
+            return {
+                "details": "note is added successfully",
+                "note": {"id": new_note.id, "title": new_note.title},
+            }
+        except Exception as e:
+            raise e
 
     async def get_user_notes(self, user_id: int) -> list[NoteOut]:
         """
@@ -132,18 +145,21 @@ class NoteService:
         :param user_id: The id of the user to get their notes.
         :return: User notes.
         """
-        notes: list[Note] | None = await self.note_repository.get_user_notes(user_id)
-        if not notes:
-            raise HTTPException(status_code=404, detail="No notes are found")
+        try:
+            notes: list[Note] | None = await self.note_repository.get_user_notes(user_id)
+            if not notes:
+                raise HTTPException(status_code=404, detail="No notes are found")
 
-        return [
-            NoteOut(
-                id=note.id,
-                title=note.title,
-                content=note.content,
-                username=note.user.username,
-                parent=ParentOut(id=note.parent.id, name=note.parent.name),
-                tags=[TagOut(id=tag.id, name=tag.name) for tag in note.tags],
-            )
-            for note in notes
-        ]
+            return [
+                NoteOut(
+                    id=note.id,
+                    title=note.title,
+                    content=note.content,
+                    username=note.user.username,
+                    parent=ParentOut(id=note.parent.id, name=note.parent.name),
+                    tags=[TagOut(id=tag.id, name=tag.name) for tag in note.tags],
+                )
+                for note in notes
+            ]
+        except Exception as e:
+            raise e
